Clean up the lock task APIs for COSU devices.
Clean up and increase readability of internal handling of lock task
mode APIs.
Add a public API to query the lock task mode state with pinned and
locked as
possible outcomes. Additionally, change wording in lock task toasts
when in
locked mode and update the javadoc regarding onLockTaskModeEntering and
onLockTaskModeExiting to represent the actual behaviour.